feat(a2a): publish shared schema package
