+++ packages/server/package.json
@@ -18,7 +18,8 @@
     "drizzle-orm": "^0.45.1",
     "hono": "^4.11.9",
     "posthog-node": "^5.29.2",
-    "zod": "^4.3.6"
+    "zod": "^4.3.6",
+    "@relaycast/a2a": "1.1.3"
   },
   "repository": {
     "type": "git",
